requirements.txt
Selenium~=4.25.0
allure-python-commons~=2.13.5